move partition mod to new crate
