+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
     "crates/package",
     "crates/runtime",
     "crates/runtime-auth",
+    "crates/runtime-table-partition",
     "crates/search",
     "crates/spice_cloud",
     "crates/spicepod",
