+++ crates/runtime-table-partition/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "runtime-table-partition"
+edition.workspace = true
+exclude.workspace = true
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+arrow-schema.workspace = true
+async-trait.workspace = true
+datafusion.workspace = true
